requirements.txt
streamlit==1.28.0
Pillow==10.1.0
rembg==2.0.52

